requirements.txt
openai
sqlalchemy
pandas
tiktoken
python-magic
trafilatura==1.5.0
langchain
fastapi
uvicorn
faiss-cpu
unstructured
python-dotenv==1.0.0
psycopg==3.1.8
pgvector==0.1.7
unstructured[local-inference]
torch
pdf2image
pinecone-client
jupyter
sentence_transformers
psycopg[binary,pool]